code improvement split the header content to other file. so that those can manage easily
