+++ src/App.js
@@ -1,28 +1,15 @@
-import ShoppingCartOutlinedIcon from '@mui/icons-material/ShoppingCartOutlined';
-import { Container } from '@mui/material';
-//MATERIAL-UI FIRSTNAVIGATIONBAR
-import Badge from '@mui/material/Badge';
 import 'bootstrap/dist/css/bootstrap.min.css';
-import React, { useEffect, useState } from 'react';
-import { useSelector } from 'react-redux';
-import {
-    BrowserRouter as Router,
-    //Routes,
-    Link,
-    Route,
-    Switch
-} from 'react-router-dom';
+import React from 'react';
+import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 // toast
 import { ToastContainer } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
 import './App.css';
-//import AuthVerify from "./common/AuthVerify";
-import EventBus from './common/EventBus';
+import Header from './components/Header/Header';
 //import FirstNavigation from "./components/layout/Header/FirstNavigation";
 import Footer from './components/layout/Footer/Footer';
 import MainNavigation from './components/layout/Header/MainNavigation';
 import Profile from './components/layout/Profile/Profile';
-import classes from './FirstNavigation.module.css';
 import Cart from './pages/Cart/Cart';
 import AccesoriosMaquillaje from './pages/Categories/AccesoriosMaquillaje';
 import AlimentosyBebidas from './pages/Categories/AlimentosyBebidas';
@@ -52,39 +39,10 @@ import ShopNow from './pages/ShopNow';
 import StorePage from './pages/Store/StorePage';
 import Stores from './pages/Stores';
 import TheNew from './pages/TheNew';
-// AUTHENTICATION
-import AuthService from './services/auth.service';
 
 const App = () => {
   // const [showModeratorBoard, setShowModeratorBoard] = useState(false);
   // const [showAdminBoard, setShowAdminBoard] = useState(false);
-  const [currentUsername, setCurrentUsername] = useState(undefined);
-  const quantity = useSelector((state) => state.cart.quantity);
-
-  useEffect(() => {
-    const user = AuthService.getCurrentUsername();
-
-    if (user) {
-      setCurrentUsername(user);
-      //  setShowModeratorBoard(user.roles.includes("ROLE_MODERATOR"));
-      //  setShowAdminBoard(user.roles.includes("ROLE_ADMIN"));
-    }
-
-    EventBus.on('logout', () => {
-      logOut();
-    });
-
-    return () => {
-      EventBus.remove('logout');
-    };
-  }, []);
-
-  const logOut = () => {
-    AuthService.logout();
-    //   setShowModeratorBoard(false);
-    //   setShowAdminBoard(false);
-    setCurrentUsername(undefined);
-  };
 
   return (
     <Router>
@@ -101,69 +59,8 @@ const App = () => {
       />
       {/*<Announcement />*/}
       {/*<FirstNavigation />*/}
-      <header className={classes.header}>
-        <Container>
-          <div className={classes.container_bar}>
-            <div className={classes.wrapper}>
-              <div className={classes.left}>
-                <div className={classes.logo}>
-                  <Link to='/'>
-                    <img
-                      src='/img/logo.png'
-                      alt='MINISO'
-                      /*height={55}
-                    width={55}*/
-                    />
-                  </Link>
-                </div>
-              </div>
-              <div className={classes.center}>
-                {/*<div className={classes.searchContainer}>
-              <input className={classes.input}></input>
-              <SearchIcon style={{ color: "gray", fontSize: 16 }} />
-  </div>*/}
-              </div>
-              <div className={classes.right}>
-                {currentUsername ? (
-                  <div className={classes.specialLinks}>
-                    <Link to={'/profile'} className={classes.link}>
-                      <div className={classes.MenuItem}>{currentUsername}</div>
-                    </Link>
-                    <div className={`nav-item ${classes.logoutButton}`}>
-                      <Link
-                        to='/login'
-                        className={classes.MenuItem}
-                        onClick={logOut}
-                      >
-                        CERRAR SESION
-                      </Link>
-                    </div>
-                  </div>
-                ) : (
-                  <div className={classes.specialLinks}>
-                    <Link to='/login' className={classes.link}>
-                      <div className={classes.MenuItem}>INICIAR SESIÓN</div>
-                    </Link>
-                    <Link to='/register' className={classes.link}>
-                      <div className={classes.MenuItem}>REGISTRARSE</div>
-                    </Link>
-                  </div>
-                )}
-                <Link to='/cart'>
-                  <div className={classes.MenuItem}>
-                    <Badge
-                      badgeContent={currentUsername ? quantity : 0}
-                      color='error'
-                    >
-                      <ShoppingCartOutlinedIcon color='primary' />
-                    </Badge>
-                  </div>
-                </Link>
-              </div>
-            </div>
-          </div>
-        </Container>
-      </header>
+
+      <Header />
       <MainNavigation />
       <Switch>
         <Route path='/' exact>
